ai-job-applier-web/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
openai
pyjwt
python-dotenv
playwright
playwright-stealth
redis
celery
tencentcloud-sdk-python
twilio
